e2e: move snapshots.spec.js tests to ariadne.spec.js loop
