+++ e2e/tests/ariadne.spec.js
@@ -13,8 +13,43 @@ for (let i = 0; i < testCasesBackendSuccess.length; i++) {
   const stubbedBackendAPIResponse = JSON.stringify(testCase.response, null, '    ');
 
   test.describe(`${testCase.name}`, () => {
+      test('HTML matches expected', async ({ page }) => {
+          //Define a mock HTTP request handler for the /v0/ URL path to intercept the request and return a mocked response.
+          await page.route('**/v0/*', async route => {
+              //Return a mock response with a JSON body and a 200 status code
+              await route.fulfill({
+                                      status: 200,
+                                      contentType: 'application/json',
+                                      body: stubbedBackendAPIResponse,
+                                  });
+          });
 
-    test('Ask a Librarian link pops up a new Ask a Library tab', async ({ page }) => {
+          await page.goto(`/${testCase.queryString}`);
+          await page.waitForSelector('h6');
+
+          const snapshot = await page.innerHTML('body');
+          // fs.writeFileSync('tests/actual/new-yorker.html', snapshot);
+          const golden = fs.readFileSync(`tests/golden/${testCase.key}.html`, { encoding: 'utf8' });
+          const ok = ( snapshot === golden );
+
+          // Actual diffing is not working in CI:
+
+          // if (!ok) {
+          //   try {
+          //     // eslint-disable-next-line no-unused-vars
+          //     execSync('diff -c tests/golden/new-yorker.html tests/actual/new-yorker.html');
+          //   } catch (error) {
+          // TODO: save stdout and stderr into variables, and print them out in the test assertions
+          //     // eslint-disable-next-line no-console
+          //     console.log(error.stdout.toString());
+          //     // eslint-disable-next-line no-console
+          //     console.log(error.stderr.toString());
+          //   }
+          // }
+          expect(ok).toBe(true);
+      });
+
+      test('Ask a Librarian link pops up a new Ask a Library tab', async ({ page }) => {
       await page.route('**/v0/*', async route => {
         await route.fulfill({
                               status: 200,
