+++ e2e/tests/golden/the-new-yorker.html
